Refactor accounts page with local components
